backend/src/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.2
django-cors-headers==4.7.0
django-mdeditor==0.1.20
django-redis==5.4.0
djangorestframework==3.16.0
idna==3.10
Markdown==3.7
pillow==11.1.0
psycopg2-binary==2.9.10
python-decouple==3.8
redis==5.2.1
requests==2.32.3
sqlparse==0.5.3
urllib3==2.4.0
whitenoise==6.9.0
